+++ src/CMakeLists.txt
@@ -28,6 +28,7 @@ add_library(llama
             llama-model.cpp
             llama-quant.cpp
             llama-sampling.cpp
+            llama-sparkinfer.cpp
             llama-vocab.cpp
             unicode-data.cpp
             unicode.cpp
