+++ src/llama-model.cpp
@@ -7,6 +7,7 @@
 #include "llama-model-loader.h"
 #include "llama-kv-cache.h"
 #include "llama.h"
+#include "llama-sparkinfer.h"
 
 #include "ggml-cpp.h"
 #include "gguf.h"
@@ -1485,804 +1486,7 @@ void llama_model::load_vocab(llama_model_loader & ml) {
     vocab.load(ml, kv);
 }
 
-/*  ----------offline split----------  */
-
-void debug_print_tensor_i64_to_file(FILE* log_file, const struct ggml_tensor* tensor) {
-    if (!log_file) {
-        return;
-    }
-
-    if (!tensor) {
-        fprintf(log_file, "debug_print_tensor_i64: tensor is NULL\n");
-        return;
-    }
-
-    // 检查类型 (这部分不变)
-    if (tensor->type != GGML_TYPE_I64) {
-        fprintf(log_file, "debug_print_tensor_i64: tensor '%s' is not INT64 (type = %d)\n", tensor->name, tensor->type);
-        return;
-    }
-
-    // 检查是否为一维
-    const int ne0 = tensor->ne[0];  // 第一维大小
-    const int ne1 = tensor->ne[1];
-    const int ne2 = tensor->ne[2];
-    const int ne3 = tensor->ne[3];
-
-    if (ne1 != 1 || ne2 != 1 || ne3 != 1) {
-        fprintf(log_file, "debug_print_tensor_i64: tensor '%s' is not 1D (shape: %d x %d x %d x %d)\n", tensor->name, ne0, ne1, ne2, ne3);
-        return;
-    }
-
-    const int64_t* data_to_print;
-    int64_t* cpu_buffer = NULL; // 用于存放从GPU拷贝回来的数据的临时缓冲区
-
-    // 检查张量是否在主机（CPU）内存中
-    // 如果 tensor->buffer 为 NULL，则它也在主机内存中
-    if (tensor->buffer == NULL || ggml_backend_buffer_is_host(tensor->buffer)) {
-        // 如果在 CPU 上，直接使用它的数据指针
-        data_to_print = (const int64_t*)tensor->data;
-    } else {
-        // 如果在 GPU 上，需要将数据拷贝回 CPU
-        const size_t tensor_size_bytes = ggml_nbytes(tensor);
-
-        // 1. 在 CPU 上分配一个临时缓冲区
-        cpu_buffer = (int64_t*)malloc(tensor_size_bytes);
-        if (!cpu_buffer) {
-            fprintf(log_file, "debug_print_tensor_i64: failed to allocate CPU buffer for tensor '%s'\n", tensor->name);
-            return;
-        }
-
-        // 2. 从 GPU (tensor->data) 拷贝数据到 CPU (cpu_buffer)
-        // ggml_backend_tensor_get 会自动处理不同后端的拷贝
-        ggml_backend_tensor_get(tensor, cpu_buffer, 0, tensor_size_bytes);
-
-        // 3. 让我们的打印指针指向这个新的 CPU 缓冲区
-        data_to_print = cpu_buffer;
-    }
-
-    if (!data_to_print) {
-        fprintf(log_file, "debug_print_tensor_i64: tensor '%s' data is NULL after backend check\n", tensor->name);
-        if (cpu_buffer) free(cpu_buffer); // 如果分配了内存，记得释放
-        return;
-    }
-
-    fprintf(log_file, "ggml_tensor name(%s) (INT64, 1D, size = %d):\n[", tensor->name, ne0);
-    for (int i = 0; i < ne0; i++) {
-        fprintf(log_file, "%" PRId64, data_to_print[i]);
-        if (i < ne0 - 1) fprintf(log_file, ", ");
-    }
-    fprintf(log_file, "]\n");
-
-    // 如果我们为 GPU 数据分配了临时缓冲区，现在就释放它
-    if (cpu_buffer) {
-        free(cpu_buffer);
-    }
-}
-
-// GTODO: need to consider no gate weight
-/**
- * @brief 管理单个层中FFN权重神经元的GPU缓存。
- * 
- * 负责在GPU上为部分卸载的神经元分配固定大小的缓存，并处理动态的换入换出操作。
- */
-struct sparkInfer_layer_cache {
-
-    ggml_backend_t gpu_backend = nullptr;
-    ggml_backend_t cpu_backend = nullptr;
-
-    // 指向CPU上的完整FFN权重张量（数据源）
-    struct ggml_tensor* cpu_ffn_gate = nullptr;
-    struct ggml_tensor* cpu_ffn_up   = nullptr;
-    struct ggml_tensor* cpu_ffn_down_t = nullptr;
-
-    // offloaded ffn neuron bufferr
-    ggml_backend_buffer_t gpu_weights_buffer = nullptr;
-    // gpu split index buffers
-    std::vector<ggml_backend_buffer_t> split_idx_allocated_buffers;
-
-    // 代表GPU缓存池中所有槽位的张量
-    struct ggml_tensor* gpu_ffn_gate_cache = nullptr;
-    struct ggml_tensor* gpu_ffn_up_cache   = nullptr;
-    struct ggml_tensor* gpu_ffn_down_t_cache = nullptr;
-
-    // ffn sparse inference relevant
-    struct ggml_tensor * ffn_gpu_neu_idx          = nullptr; // on gpu
-    struct ggml_tensor * ffn_gpu_neu_mask         = nullptr;
-    struct ggml_tensor * ffn_gpu_group_idx        = nullptr; // on gpu
-    struct ggml_tensor * ffn_gpu_group_mask       = nullptr;
-    struct ggml_tensor * ffn_neuron_to_group_map  = nullptr;
-
-    int64_t neuron_cache_capacity = 0; // GPU缓存池能容纳的神经元数量
-    uint64_t layer_neuron_count = 0; // 每层神经元总量
-    uint64_t layer_group_count = 0; // 每层分组总量
-    uint64_t layer_group_size = 0; // 每层分组大小
-    
-    // 跟踪: 原始神经元索引 -> GPU缓存槽位索引
-    std::unordered_map<int64_t, int64_t> neuron_to_slot_map;
-    
-    // 跟踪: GPU缓存槽位索引 -> 原始神经元索引
-    std::vector<int64_t> slot_to_neuron_map;
-
-    // 实现LRU(最近最少使用)的替换策略，存储的是【原始神经元索引】
-    std::list<int64_t> lru_tracker;
-    std::unordered_map<int64_t, std::list<int64_t>::iterator> lru_map;
-
-    // 临时的ggml上下文，用于创建视图等临时张量
-    struct ggml_context* tmp_ctx = nullptr;
-
-    size_t offloaded_bytes=0;
-
-    ~sparkInfer_layer_cache() {
-        if (gpu_weights_buffer) {
-            ggml_backend_buffer_free(gpu_weights_buffer);
-        }
-        if (tmp_ctx) {
-            ggml_free(tmp_ctx);
-        }
-    }
-
-    /**
-     * @brief 初始化层的缓存。
-     * 
-     * @param layer llama模型中的层。
-     * @param backend 用于GPU操作的后端。
-     * @param initial_gpu_neuron_indices 初始需要加载到GPU的神经元原始索引列表。
-     * @return true 如果初始化成功。
-     */
-    bool init(int layer_idx, llama_model& model, llama_layer& layer, ggml_backend_t backend, const std::vector<int64_t>& initial_gpu_neuron_indices) {
-        // init
-        this->gpu_backend = backend;
-        this->cpu_backend = ggml_backend_cpu_init();
-
-        bool has_gate = true; // GTODO: gate diverage
-        bool full_gpu = layer.gpu_offload_ratio >= 1;
-
-        this->cpu_ffn_gate   = has_gate ? layer.ffn_gate : nullptr;
-        this->cpu_ffn_up     = layer.ffn_up;
-        this->cpu_ffn_down_t = layer.ffn_down_t;
-        
-        this->layer_neuron_count = model.layer_neuron_count; // 每层神经元总数
-        this->layer_group_size = model.layer_group_size; // 每层分组大小
-        this->layer_group_count = model.layer_group_count; // 每层分组数量
-        this->neuron_cache_capacity = initial_gpu_neuron_indices.size();
-        if (this->neuron_cache_capacity == 0) return true; // 无需卸载
-
-        ffn_gpu_neu_idx = layer.ffn_gpu_neu_idx;
-        ffn_gpu_neu_mask = layer.ffn_gpu_neu_mask;
-        ffn_gpu_group_idx = layer.ffn_gpu_group_idx;
-        ffn_gpu_group_mask = layer.ffn_gpu_group_mask;
-        ffn_neuron_to_group_map = layer.ffn_neuron_to_group_map;
-
-        /* debug info */ 
-        // FILE* log_file = fopen("debug_split_info.log", "a");
-        // if (log_file == NULL) {
-        //     // 如果文件打开失败，可以打印一个错误到 stderr 然后继续，或者直接退出
-        //     perror("Failed to open log file");
-        //     // return; // 或者根据你的错误处理逻辑决定是否返回
-        // }
-        // std::time_t result = std::time(nullptr);
-        // fprintf(log_file, "\n--- Debugging Layer %d split info into file, timestamp %s ---", layer_idx, std::ctime(&result)); // 添加一些上下文信息
-        // debug_print_tensor_i64_to_file(log_file, ffn_gpu_neu_idx);
-        // debug_print_tensor_i64_to_file(log_file, ffn_gpu_neu_mask);
-        // debug_print_tensor_i64_to_file(log_file, ffn_gpu_group_idx);
-        // debug_print_tensor_i64_to_file(log_file, ffn_gpu_group_mask);
-        // debug_print_tensor_i64_to_file(log_file, ffn_neuron_to_group_map);
-        // fflush(log_file);
-        // fclose(log_file);
-        /* debug info end */ 
-
-        GGML_ASSERT(neuron_cache_capacity <= layer_neuron_count && "we required neuron_cache_capacity <= layer_neuron_count");
-
-        // 1. 计算并分配ffn buffer size
-        const size_t single_mat_size = ggml_backend_buft_get_alloc_size(
-            ggml_backend_get_default_buffer_type(gpu_backend),
-            cpu_ffn_down_t // 使用其中一个矩阵作为尺寸参考
-        );
-        // 我们只缓存部分行，所以要按比例计算
-        const size_t single_cache_size = (single_mat_size / layer_neuron_count) * neuron_cache_capacity;
-        
-        // 总大小 = 3个矩阵的缓存大小之和
-        const size_t total_gpu_buffer_size = single_cache_size * 3;
-
-        this->gpu_weights_buffer = ggml_backend_alloc_buffer(gpu_backend, total_gpu_buffer_size);
-        if (!gpu_weights_buffer) {
-            LLAMA_LOG_ERROR("%s: failed to allocate GPU buffer for layer cache\n", __func__);
-            return false;
-        }
-
-        // 2. 在GPU缓存池中创建代表缓存张量
-        struct ggml_init_params params = { ggml_tensor_overhead() * 6, NULL, true };
-        this->tmp_ctx = ggml_init(params);
-
-        void* current_addr = ggml_backend_buffer_get_base(gpu_weights_buffer);
-        
-        char gate_name[64];
-
-        if(has_gate) {
-            gpu_ffn_gate_cache = ggml_new_tensor_2d(tmp_ctx, cpu_ffn_gate->type, cpu_ffn_gate->ne[0], neuron_cache_capacity);
-            snprintf(gate_name, sizeof(gate_name), "blk.%d.ffn_qgu_gate.weight", layer_idx);
-            ggml_set_name(gpu_ffn_gate_cache, gate_name);
-            ggml_backend_tensor_alloc(gpu_weights_buffer, gpu_ffn_gate_cache, current_addr);
-            current_addr = (char*)current_addr + single_cache_size;
-        }
-
-        gpu_ffn_up_cache = ggml_new_tensor_2d(tmp_ctx, cpu_ffn_up->type, cpu_ffn_up->ne[0], neuron_cache_capacity);
-        snprintf(gate_name, sizeof(gate_name), "blk.%d.ffn_qgu_up.weight", layer_idx);
-        ggml_set_name(gpu_ffn_up_cache, gate_name);
-        ggml_backend_tensor_alloc(gpu_weights_buffer, gpu_ffn_up_cache, current_addr);
-        current_addr = (char*)current_addr + single_cache_size;
-
-        gpu_ffn_down_t_cache = ggml_new_tensor_2d(tmp_ctx, cpu_ffn_down_t->type, cpu_ffn_down_t->ne[0], neuron_cache_capacity);
-        snprintf(gate_name, sizeof(gate_name), "blk.%d.ffn_qgu_down_t.weight", layer_idx);
-        ggml_set_name(gpu_ffn_down_t_cache, gate_name);
-        ggml_backend_tensor_alloc(gpu_weights_buffer, gpu_ffn_down_t_cache, current_addr);
-        
-        // 将新的GPU缓存张量赋给llama_layer
-        layer.ffn_gpu_gate = has_gate ? gpu_ffn_gate_cache : nullptr;
-        layer.ffn_gpu_up   = gpu_ffn_up_cache;
-        layer.ffn_gpu_down_t = gpu_ffn_down_t_cache;
-
-        // 3. init slot_to_neuron_map from ffn_gpu_neu_idx, and copy data to buffer
-        auto t_start = ggml_time_ms();
-        this->slot_to_neuron_map.resize(neuron_cache_capacity, -1);
-
-        auto batch_copy_neurons = [&](ggml_tensor* cpu_src, ggml_tensor* gpu_dst_cache, const std::vector<int64_t>& indices, const bool full_gpu) {
-            if(full_gpu){
-                const size_t full_tensor_bytes = ggml_nbytes(cpu_src);
-                ggml_backend_tensor_set(gpu_dst_cache, cpu_src->data, 0, full_tensor_bytes);
-            }else{
-                const int64_t n_embd = cpu_src->ne[0];
-                const size_t row_size_bytes = ggml_row_size(cpu_src->type, n_embd);
-                
-                // 在CPU上分配一个临时暂存缓冲区
-                std::vector<char> staging_buffer(row_size_bytes * indices.size());
-                
-                for (size_t i = 0; i < indices.size(); ++i) {
-                    const int64_t neuron_idx = indices[i];
-                    
-                    // 源地址：在完整CPU张量中的位置
-                    char* src_ptr = (char*)cpu_src->data + neuron_idx * cpu_src->nb[1];
-                    
-                    // 目标地址：在暂存缓冲区中的位置
-                    char* dst_ptr = staging_buffer.data() + i * row_size_bytes;
-                    
-                    memcpy(dst_ptr, src_ptr, row_size_bytes);
-                }
-                
-                ggml_backend_tensor_set(gpu_dst_cache, staging_buffer.data(), 0, staging_buffer.size());
-            }
-        };
-
-        if(has_gate) batch_copy_neurons(cpu_ffn_gate, gpu_ffn_gate_cache, initial_gpu_neuron_indices, full_gpu);
-        batch_copy_neurons(cpu_ffn_up, gpu_ffn_up_cache, initial_gpu_neuron_indices, full_gpu);
-        batch_copy_neurons(cpu_ffn_down_t, gpu_ffn_down_t_cache, initial_gpu_neuron_indices, full_gpu);
-
-        // 更新元数据
-        offloaded_bytes += ggml_nbytes(gpu_ffn_up_cache) * (has_gate ? 3 : 2); // 每个神经元有3个矩阵
-        for (size_t i = 0; i < initial_gpu_neuron_indices.size(); ++i) {
-            int64_t neuron_idx = initial_gpu_neuron_indices[i];
-            int64_t slot_idx = i;
-            update_metadata(neuron_idx, slot_idx);
-        }
-
-        auto t_end = ggml_time_ms();
-        LLAMA_LOG_INFO("%s: layer %d offload in %lld ms, cached %d neurons %s\n", __func__, layer_idx, t_end - t_start, neuron_cache_capacity, full_gpu?"(full_gpu)":" ");
-
-        return true;
-    }
-
-    /**
-     * @brief 确保一个神经元在GPU上。如果不在，就执行换入操作。
-     * 
-     * @param neuron_idx 神经元的原始索引。
-     * @return int32_t 神经元在GPU缓存中的槽位索引。
-     */
-    int64_t ensure_neuron_on_gpu(int64_t neuron_idx) {
-        auto it = neuron_to_slot_map.find(neuron_idx);
-        if (it != neuron_to_slot_map.end()) {
-            // Cache Hit: 神经元已在GPU上，更新其为最近使用
-            int64_t slot_idx = it->second;
-            lru_tracker.erase(lru_map[neuron_idx]);
-            lru_tracker.push_front(neuron_idx);
-            lru_map[neuron_idx] = lru_tracker.begin();
-            return slot_idx;
-        }
-
-        // Cache Miss: 需要换入
-        int64_t slot_to_use = -1;
-        if (neuron_to_slot_map.size() < neuron_cache_capacity) {
-            // 缓存未满，找到一个空槽位
-            for(int64_t i = 0; i < (int64_t)slot_to_neuron_map.size(); ++i) {
-                if (slot_to_neuron_map[i] == -1) {
-                    slot_to_use = i;
-                    break;
-                }
-            }
-        } else {
-            // 缓存已满，根据LRU策略选择牺牲品
-            int64_t victim_neuron_idx = lru_tracker.back();
-            lru_tracker.pop_back();
-            
-            slot_to_use = neuron_to_slot_map[victim_neuron_idx];
-            
-            neuron_to_slot_map.erase(victim_neuron_idx);
-            lru_map.erase(victim_neuron_idx);
-        }
-
-        GGML_ASSERT(slot_to_use != -1);
-
-        // 执行拷贝并将新神经元信息写入元数据
-        copy_neuron_to_gpu_slot(neuron_idx, slot_to_use);
-        update_metadata(neuron_idx, slot_to_use);
-        
-        return slot_to_use;
-    }
-
-private:
-    /**
-     * @brief 将一个神经元从CPU复制到GPU的指定slot。
-     */
-    void copy_neuron_to_gpu_slot(int64_t neuron_idx, int64_t slot_idx) {
-        // 创建一个临时的上下文用于视图操作
-        struct ggml_init_params params = { ggml_tensor_overhead() * 6, NULL, true };
-        struct ggml_context* view_ctx = ggml_init(params);
-
-
-        auto copy_row = [&](ggml_tensor* cpu_src, ggml_tensor* gpu_dst_cache) {
-            const int64_t row_len = cpu_src->ne[0];
-            const size_t  row_bytes = ggml_row_size(cpu_src->type, row_len);
-
-            // 创建指向CPU源矩阵中特定行的视图
-            struct ggml_tensor* src_view = ggml_view_1d(view_ctx, cpu_src, row_len, neuron_idx * cpu_src->nb[1]);
-            src_view->buffer = cpu_src->buffer;
-
-            // 创建指向GPU缓存中特定槽位的视图
-            struct ggml_tensor* dst_view = ggml_view_1d(view_ctx, gpu_dst_cache, row_len, slot_idx * gpu_dst_cache->nb[1]);
-            dst_view->buffer = gpu_dst_cache->buffer;
-
-            // 使用后端进行异步拷贝
-            ggml_backend_tensor_copy(src_view, dst_view);
-        };
-
-        copy_row(cpu_ffn_gate, gpu_ffn_gate_cache);
-        copy_row(cpu_ffn_up,   gpu_ffn_up_cache);
-        copy_row(cpu_ffn_down_t, gpu_ffn_down_t_cache);
-        
-        ggml_free(view_ctx);
-    }
-    
-    /**
-     * @brief 更新缓存的元数据。
-     */
-    void update_metadata(int64_t neuron_idx, int64_t slot_idx) {
-        neuron_to_slot_map[neuron_idx] = slot_idx;
-        slot_to_neuron_map[slot_idx] = neuron_idx;
-        lru_tracker.push_front(neuron_idx);
-        lru_map[neuron_idx] = lru_tracker.begin();
-    }
-};
-
-
-/**
- * @brief 全局神经元缓存管理器，管理模型中所有层的缓存。
- *
- */
-struct sparkInfer_neuron_cache_manager {
-    std::vector<sparkInfer_layer_cache> layer_caches;
-    llama_model *model = nullptr;
-    size_t total_offloaded_bytes=0;
-
-    bool init(llama_model &p_model, ggml_backend_t gpu_backend) {
-        this->model = &p_model;
-        const int n_layers = model->hparams.n_layer;
-        layer_caches.resize(n_layers);
-
-        const int64_t t_start_us = ggml_time_us();
-        LLAMA_LOG_INFO("%s: Initializing neuron cache for %d layers...\n", __func__, n_layers);
-
-        // #pragma omp parallel for
-        for (int i = 0; i < n_layers; ++i) {
-            llama_layer &layer = model->layers[i];
-            if (layer.gpu_offload_ratio == 0.f || !layer.ffn_gpu_neu_idx) {
-                LLAMA_LOG_INFO("%s: layer %d won't split partial tensor to GPU", __func__, i);
-                continue; // 该层不卸载或没有索引信息
-            }
-
-            // 从ffn_gpu_neu_idx读取需要加载的神经元
-            struct ggml_tensor* initial_indices_tensor = layer.ffn_gpu_neu_idx;
-            std::vector<int64_t> initial_indices(initial_indices_tensor->ne[0]);
-            
-            // 注意：这里需要从设备或主机内存中获取数据 TAG
-            ggml_backend_tensor_get(initial_indices_tensor, initial_indices.data(), 0, ggml_nbytes(initial_indices_tensor));
-            
-            if (!layer_caches[i].init(i ,p_model, layer, gpu_backend, initial_indices)) {
-                LLAMA_LOG_ERROR("%s: failed to initialize cache for layer %d\n", __func__, i);
-                throw std::runtime_error("Failed to initialize layer cache");
-            }
-            total_offloaded_bytes += layer_caches[i].offloaded_bytes;
-        }
-        
-        // 在所有层初始化后，可以释放掉模型中用于加载的静态索引缓冲区
-        // for (auto& buffer : model->split_idx_allocated_buffers) {
-        //     ggml_backend_buffer_free(buffer);
-        // }
-        // model->split_idx_allocated_buffers.clear();
-        // if (model->ctx_cpu_idx_tensors) {
-        //     ggml_free(model->ctx_cpu_idx_tensors);
-        //     model->ctx_cpu_idx_tensors = nullptr;
-        // }
-        // if (model->ctx_gpu_idx_tensors) {
-        //     ggml_free(model->ctx_gpu_idx_tensors);
-        //     model->ctx_gpu_idx_tensors = nullptr;
-        // }
-
-
-        const int64_t t_end_us = ggml_time_us();
-        LLAMA_LOG_INFO("%s: Neuron cache initialized successfully. (%.2f ms)\n", __func__, (t_end_us - t_start_us) / 1000.0);
-
-        return true;
-    }
-
-    /**
-     * @brief 在推理前，准备好指定层所需的一组神经元。
-     *
-     * @param layer_idx 层的索引。
-     * @param required_neuron_indices 需要确保在GPU上的神经元原始索引列表。
-     * @param out_gpu_slot_indices [输出参数] 填充更新后的GPU槽位索引，用于计算图。
-     */
-    void prepare_hot_neurons(int layer_idx, const std::vector<int64_t>& required_neuron_indices, std::vector<int64_t>& out_gpu_slot_indices) {
-        if (layer_caches[layer_idx].neuron_cache_capacity == 0) return;
-
-        out_gpu_slot_indices.resize(required_neuron_indices.size());
-        for (size_t i = 0; i < required_neuron_indices.size(); ++i) {
-            out_gpu_slot_indices[i] = layer_caches[layer_idx].ensure_neuron_on_gpu(required_neuron_indices[i]);
-        }
-
-        // 实际应用中，你需要用 out_gpu_slot_indices 的数据更新计算图中使用的索引张量。
-        // 例如：
-        // struct ggml_tensor* graph_neu_idx = model->layers[layer_idx].ffn_gpu_neu_idx;
-        // ggml_backend_tensor_set(graph_neu_idx, out_gpu_slot_indices.data(), 0, sizeof(int32_t) * out_gpu_slot_indices.size());
-    }
-};
-
-/**
- * @brief init，用于初始化并管理FFN缓存。
- * 
- */
-sparkInfer_neuron_cache_manager* sparkinfer_init_and_manage_ffn_cache(struct llama_model* model, ggml_backend_t gpu_backend) {
-    if (!model || !gpu_backend) {
-        throw std::invalid_argument("sparkinfer_init_and_manage_ffn_cache: model or gpu_backend is null");
-    }
-
-    auto cache_manager = new sparkInfer_neuron_cache_manager();
-    if (!cache_manager->init(*model, gpu_backend)) {
-        throw std::runtime_error("sparkinfer_init_and_manage_ffn_cache: failed to initialize cache manager");
-    }
-
-    return cache_manager;
-}
-
-// static int64_t sum_gpu_index(struct ggml_tensor * gpu_index) {
-//     ggml_context * ctx_aux = ggml_init({
-//         /* mem_size */ 1 << 10,
-//     });
-
-//     GGML_ASSERT(ctx_aux);
-
-//     ggml_cgraph * gf = ggml_new_graph_custom(ctx_aux, 1, false);
-//     ggml_tensor * sum = ggml_sum(ctx_aux, gpu_index);
-
-//     ggml_set_name(sum, "gpu_index_sum");
-//     ggml_build_forward_expand(gf, sum);
-
-//     // TODO: +1 worker for GPU under hybrid inference but no use
-//     // ggml_graph_compute_helper(work_buffer, gf, 2);
-//     ggml_graph_compute_with_ctx(ctx_aux, gf, 2);
-
-//     int32_t sum_val = ggml_get_i32_1d(sum, 0);
-
-//     ggml_free(ctx_aux);
-
-//     return sum_val;
-// }
-
-struct sparkinfer_split_loader {
-    std::string fname;
-    struct gguf_context * ctx_gguf = nullptr;
-    struct ggml_context * ctx_meta = nullptr;
-
-    std::vector<ggml_backend_buffer_t> allocated_buffers;
-    // 分别管理CPU和GPU张量的元数据
-    struct ggml_context * ctx_cpu_tensors = nullptr;
-    struct ggml_context * ctx_gpu_tensors = nullptr;
-
-    int n_tensors = 0;
-    uint64_t vram_required = 0;
-    uint64_t layer_neuron_count = 0;
-    uint64_t layer_group_count = 0;
-    uint64_t layer_group_size = 0;
-
-    sparkinfer_split_loader(const std::string & fname) : fname(fname) {
-        struct gguf_init_params params = {
-            /* .no_alloc = */ false,
-            /* .ctx      = */ &ctx_meta,
-        };
-
-        ctx_gguf = gguf_init_from_file(fname.c_str(), params);
-        if (!ctx_gguf) {
-            throw std::runtime_error("无法打开或加载分割文件: " + fname);
-        }
-
-        // 读取元数据
-        int key_idx = gguf_find_key(ctx_gguf, "split.vram_capacity");
-        if (key_idx >= 0) {
-            vram_required = gguf_get_val_u64(ctx_gguf, key_idx);
-        } else {
-             LLAMA_LOG_WARN("警告: 'split.vram_capacity' key not found in %s\n", fname.c_str());
-        }
-
-        key_idx = gguf_find_key(ctx_gguf, "split.layer_neuron_count");
-        if (key_idx >= 0) {
-            layer_neuron_count = gguf_get_val_u64(ctx_gguf, key_idx);
-        } else {
-             LLAMA_LOG_WARN("警告: 'split.layer_neuron_count' key not found in %s\n", fname.c_str());
-        }
-
-        key_idx = gguf_find_key(ctx_gguf, "split.layer_group_count");
-        if (key_idx >= 0) {
-            layer_group_count = gguf_get_val_u64(ctx_gguf, key_idx);
-        } else {
-             LLAMA_LOG_WARN("警告: 'split.layer_group_count' key not found in %s\n", fname.c_str());
-        }
-
-        layer_group_size = layer_neuron_count / layer_group_count;
-        n_tensors = gguf_get_n_tensors(ctx_gguf);
-        
-        LLAMA_LOG_INFO("%s: 成功加载分割文件 '%s' 的信息和数据. 张量数量: %d, VRAM需求: %zu 字节\n",
-            __func__, fname.c_str(), n_tensors, (size_t)vram_required);
-    }
-
-    ~sparkinfer_split_loader() {
-        if (ctx_gguf) gguf_free(ctx_gguf);
-        if (ctx_meta) ggml_free(ctx_meta); // ctx_meta 现在管理着数据，ggml_free会释放它
-        // ctx_cpu_tensors 和 ctx_gpu_tensors 会被模型接管，由模型管理其生命周期
-        
-        // 分配的缓冲区也应由模型管理
-        // for (auto & buffer : allocated_buffers) {
-        //     ggml_backend_buffer_free(buffer);
-        // }
-    }
-
-    /**
-     * @brief 加载分割张量并将其应用到模型层。
-     *        为动态张量预分配最大容量。
-     * 
-     * @param model 要应用张量的llama模型。
-     * @param gpu_backend 用于卸载的GPU后端句柄，如果只想用CPU则为nullptr。
-     * @return true 如果成功。
-     */
-    bool load_and_apply_split(llama_model & model, ggml_backend_t gpu_backend) {
-        const int n_layers = model.hparams.n_layer;
-
-        if (n_tensors != n_layers * 5) {
-           LLAMA_LOG_ERROR("%s: 错误: 分割文件中的张量数量 (%d) 与模型层数 (%d * 5) 不匹配\n",
-                __func__, n_tensors, n_layers);
-            return false;
-        }
-        if (layer_neuron_count <= 0) {
-            throw std::runtime_error("layer_neuron_count 必须为正数");
-        }
-
-        const int64_t t_start_us = ggml_time_us();
-
-        // 1. 初始化用于管理 CPU 和 GPU 张量元数据的上下文
-        struct ggml_init_params params = { ggml_tensor_overhead() * n_tensors, NULL, true };
-        ctx_cpu_tensors = ggml_init(params);
-        ctx_gpu_tensors = ggml_init(params);
-        if (!ctx_cpu_tensors || !ctx_gpu_tensors) {
-            throw std::runtime_error("无法初始化CPU/GPU张量上下文");
-        }
-
-        // 2. 遍历层，创建目标张量元数据并分配到正确的上下文中
-        for (int il = 0; il < n_layers; ++il) {
-            llama_layer & layer = model.layers[il];
-            
-            struct ggml_tensor* temp_neu_idx_meta = get_tensor_meta_from_gguf(il, "ffn_gpu_neu_idx");
-            if (!temp_neu_idx_meta) throw std::runtime_error("无法从GGUF获取 ffn_gpu_neu_idx 元数据");
-
-            int64_t initial_gpu_neurons = temp_neu_idx_meta->ne[0];
-
-            bool offload_layer = gpu_backend && initial_gpu_neurons > 0;
-            layer.gpu_offload_ratio = float(initial_gpu_neurons) / float(layer_neuron_count);
-            ggml_context* target_ctx = offload_layer ? ctx_gpu_tensors : ctx_cpu_tensors;
-
-            LLAMA_LOG_INFO("%s: layer %2d: offload %.2f%% to GPU\n", __func__, il, layer.gpu_offload_ratio * 100.0f);
-
-            // --- 在目标上下文中创建张量元数据 ---
-            layer.ffn_gpu_neu_idx = create_static_tensor_in_ctx(target_ctx, il, "ffn_gpu_neu_idx");
-            layer.ffn_gpu_group_idx = create_static_tensor_in_ctx(target_ctx, il, "ffn_gpu_group_idx");
-            layer.ffn_gpu_neu_mask = create_static_tensor_in_ctx(ctx_cpu_tensors, il, "ffn_gpu_neu_mask");
-            layer.ffn_gpu_group_mask = create_static_tensor_in_ctx(ctx_cpu_tensors, il, "ffn_gpu_group_mask");
-            layer.ffn_neuron_to_group_map = create_static_tensor_in_ctx(ctx_cpu_tensors, il, "ffn_neuron_to_group_map");
-
-            if (!layer.ffn_gpu_neu_idx || !layer.ffn_gpu_group_idx || !layer.ffn_gpu_neu_mask || !layer.ffn_gpu_group_mask || !layer.ffn_neuron_to_group_map) {
-                throw std::runtime_error("在目标上下文中创建张量元数据失败，层 " + std::to_string(il));
-            }
-        }
-        
-        // 3. 为CPU和GPU上下文批量分配缓冲区
-        ggml_backend_t cpu_backend = ggml_backend_cpu_init();
-        if (ggml_get_first_tensor(ctx_cpu_tensors)) {
-            ggml_backend_buffer_t cpu_buffer = ggml_backend_alloc_ctx_tensors(ctx_cpu_tensors, cpu_backend);
-            if (!cpu_buffer) {
-                ggml_backend_free(cpu_backend);
-                throw std::runtime_error("为CPU张量分配缓冲区失败");
-            }
-            allocated_buffers.push_back(cpu_buffer);
-        }
-
-        if (gpu_backend && ggml_get_first_tensor(ctx_gpu_tensors)) {
-            ggml_backend_buffer_t gpu_buffer = ggml_backend_alloc_ctx_tensors(ctx_gpu_tensors, gpu_backend);
-            if (!gpu_buffer) {
-                ggml_backend_free(cpu_backend); // cpu_backend 依然需要释放
-                throw std::runtime_error("为GPU张量分配缓冲区失败");
-            }
-            allocated_buffers.push_back(gpu_buffer);
-        }
-
-        // 4. 修改点: 使用 ggml_backend_tensor_set 从已加载数据的源张量设置目标张量
-        for (int i = 0; i < gguf_get_n_tensors(ctx_gguf); ++i) {
-            const char* name = gguf_get_tensor_name(ctx_gguf, i);
-            
-            // 源张量 (数据已由 gguf_init 加载到其 ->data 指针中)
-            struct ggml_tensor* src_tensor = ggml_get_tensor(ctx_meta, name);
-            if (!src_tensor) {
-                 throw std::runtime_error("在源 GGUF 上下文中未找到张量: " + std::string(name));
-            }
-            if (!src_tensor->data) {
-                throw std::runtime_error("源张量数据未加载(data is null): " + std::string(name));
-            }
-
-            // 查找目标张量 (已由后端分配好缓冲区)
-            struct ggml_tensor* dst_tensor = ggml_get_tensor(ctx_cpu_tensors, name);
-            if (!dst_tensor) {
-                dst_tensor = ggml_get_tensor(ctx_gpu_tensors, name);
-            }
-
-            if (dst_tensor) {
-                // 使用 ggml_backend_tensor_set 将数据从 host (src_tensor->data) 复制到 device (dst_tensor)
-                // 这会自动处理 CPU->CPU 或 CPU->GPU 的情况
-                const size_t nbytes = ggml_nbytes(src_tensor);
-                ggml_backend_tensor_set(dst_tensor, src_tensor->data, 0, nbytes);
-            } else {
-                throw std::runtime_error("在目标上下文中未找到张量: " + std::string(name));
-            }
-        }
-
-        const int64_t t_end_us = ggml_time_us();
-        LLAMA_LOG_INFO("%s: 分割数据复制完毕，耗时 %.2f 毫秒\n", __func__, (t_end_us - t_start_us) / 1000.0);
-        
-
-        // 5. 保存数据到 llama_model (逻辑不变)
-        model.ctx_cpu_idx_tensors = ctx_cpu_tensors;
-        model.ctx_gpu_idx_tensors = ctx_gpu_tensors;
-
-        model.layer_neuron_count = layer_neuron_count;
-        model.layer_group_count = layer_group_count;
-        model.layer_group_size = layer_group_size;
-        model.split_idx_allocated_buffers = allocated_buffers;
-
-        return true;
-    }
-
-private:
-    struct ggml_tensor* get_tensor_meta_from_gguf(int layer_idx, const std::string& suffix) {
-        char name[GGML_MAX_NAME];
-        snprintf(name, sizeof(name), "blk.%d.%s", layer_idx, suffix.c_str());
-        return ggml_get_tensor(ctx_meta, name);
-    }
-
-    struct ggml_tensor* create_static_tensor_in_ctx(ggml_context* ctx, int layer_idx, const std::string& suffix) {
-        struct ggml_tensor* t_meta = get_tensor_meta_from_gguf(layer_idx, suffix.c_str());
-        if (!t_meta) {
-            LLAMA_LOG_WARN("%s: 无法找到张量元数据 blk.%d.%s\n", __func__, layer_idx, suffix.c_str());
-            return nullptr;
-        }
-        struct ggml_tensor* new_t = ggml_dup_tensor(ctx, t_meta);
-        ggml_set_name(new_t, t_meta->name);
-        return new_t;
-    }
-};
-
-static bool sparkinfer_load_gpu_split_from_split_file(llama_model & model, std::string split_path, size_t vram_allocatable_bytes) {
-
-    sparkinfer_split_loader loader(split_path);
-    
-    ggml_backend_t gpu_backend = ggml_backend_cuda_init(0); 
-    if (!gpu_backend) {
-        fprintf(stderr, "No GPU backend found\n");
-        return false;
-    }
-    return loader.load_and_apply_split(model,gpu_backend);
-}
-
-static bool sparkinfer_load_gpu_split_with_budget(llama_model_loader & ml, llama_model & model, size_t vram_allocatable_bytes, bool no_cache) {
-    std::string cached_split_path = ml.fname + ".sparkinfer_split_idx";//"/share/models/prosparse-7b-gguf-w-our-predictor/prosparse-7b.gguf.sparkinfer_split_idx"
-    // get basedir from cached split path
-    std::string model_basedir;
-    if (cached_split_path.find_last_of('/') != std::string::npos) {
-        model_basedir = cached_split_path.substr(0, cached_split_path.find_last_of('/'));
-    }
-
-    if (no_cache) {
-        LLAMA_LOG_INFO("%s: no cache, will generate gpu split\n", __func__);
-    } else {
-        LLAMA_LOG_INFO("%s: loading gpu split from '%s'\n", __func__, cached_split_path.c_str());
-    }
-
-    // Load GPU split from previously generated cache
-    if (access(cached_split_path.c_str(), F_OK) == 0 && !no_cache) {
-        if (sparkinfer_load_gpu_split_from_split_file(model, cached_split_path, vram_allocatable_bytes)) {
-            return true;
-        }
-        LLAMA_LOG_ERROR("%s: error: failed to apply previously generated gpu split from '%s'\n", __func__, cached_split_path.c_str());
-        throw std::runtime_error("Failed to load GPU split from cache");
-    }
-
-    // Generate GPU split
-    std::string activation_path = std::string(model_basedir)+"/activation";    
-
-    // Calculate solver parameters
-    ggml_tensor * ffn_up = model.layers[0].ffn_up;
-    ggml_tensor * ffn_gate = model.layers[0].ffn_gate;
-    int slice_size = ffn_up->ne[1] * ggml_type_size(ffn_up->type) / ggml_blck_size(ffn_up->type);
-    // For model arch with FFN gate, the gate is also sliced, otherwise only the up and down matrices are sliced
-    int vram_bytes_per_slice = slice_size * (ffn_gate ? 3 : 2); // GTODO: 4.5 or 3?
-    int neuron_cap = floor((double)vram_allocatable_bytes / vram_bytes_per_slice) * 4;
-
-    LLAMA_LOG_INFO("invoking sparkinfer Python module to generate gpu split for %.2f MiB of VRAM\n", vram_allocatable_bytes / 1024.0 / 1024.0);
-
-    std::stringstream command_ss;
-    command_ss << "python -m sparkinfer"
-               << " --model-path " << model_basedir
-               << " --layer " << model.hparams.n_layer
-               << " --neuron " << ffn_up->ne[1]
-               << " --neuron-capacity " << neuron_cap
-               << " --vram-capacity " << vram_allocatable_bytes
-               << " --output " << cached_split_path;
-    if (system(command_ss.str().c_str()) != 0 || access(cached_split_path.c_str(), F_OK) != 0) {
-        LLAMA_LOG_ERROR("%s: error: failed to generate gpu split\n", __func__);
-        return false;
-    }
-
-    return sparkinfer_load_gpu_split_from_split_file(model, cached_split_path, vram_allocatable_bytes);
-}
-
-static size_t sparkinfer_load_gpu_split_and_offload_weight(llama_model_loader & ml, llama_model & model, size_t vram_budget_bytes,bool no_cache) {
-    // log the VRAM budget
-    LLAMA_LOG_INFO("%s: VRAM budget is %.2f MiB\n", __func__, vram_budget_bytes / 1024.0 / 1024.0);
-
-    if (!sparkinfer_load_gpu_split_with_budget(ml, model, vram_budget_bytes, no_cache)) {
-        LLAMA_LOG_ERROR("%s: error: failed to generate gpu split, an empty one will be used\n", __func__);
-    }
-
-    ggml_backend_t gpu_backend = ggml_backend_cuda_init(0); 
-    if (!gpu_backend) {
-        fprintf(stderr, "No GPU backend found\n");
-        return false;
-    }
-    sparkInfer_neuron_cache_manager* neuron_cache_manager = sparkinfer_init_and_manage_ffn_cache(&model, gpu_backend);
-    if (!neuron_cache_manager) {
-        throw std::runtime_error("Failed to initialize neuron cache manager");
-    }
-
-    model.neuron_cache_manager = neuron_cache_manager;
-    size_t total_offloaded_bytes = neuron_cache_manager->total_offloaded_bytes;
-    LLAMA_LOG_INFO("%s: offloaded %.2f MiB of FFN weights to GPU\n", __func__, total_offloaded_bytes / 1024.0 / 1024.0);
-
-    return total_offloaded_bytes;
-}
+/*-------- sparse laoding --------*/
 
 static bool sparkinfer_reduce_vram_budget(size_t &vram_budget_bytes, size_t reduce_bytes) {
 
@@ -2294,7 +1498,6 @@ static bool sparkinfer_reduce_vram_budget(size_t &vram_budget_bytes, size_t redu
     return false;
 }
 
-/*  ----------offline split----------  */
 
 bool llama_model::load_sparse_tensors(llama_model_loader &ml, size_t vram_budget_bytes) {
     const auto & split_mode   = params.split_mode;
@@ -2889,6 +2092,8 @@ bool llama_model::load_sparse_tensors(llama_model_loader &ml, size_t vram_budget
     return true;
     }
 }
+
+/*-------- sparse laoding --------*/
 
 
 bool llama_model::load_tensors(llama_model_loader & ml) {
